ci: check MSRV, add a rust-toolchain file
This adds a CI job that will check the MSRV of any PR is set correctly on a crate by crate basis.
If the MSRV is set in correctly, due to e.g. some feature used, it will compute the correct MSRV for the user and indicate it in its error.
The script's error message and functioning can be emulated by manually changing the MSRV to an incorrect version (e.g. 1.87) and running it (./scripts/check-msrv.sh).

This further adds a rust-toolchain file that will make CI follow the latest stable as it updates.

As a consequence, the following behaviors can change from being proactive to being on-demand:
- bumping the MSRV.

As a side effect, we no longer have to bump our MSRV beyond what we actually use, i.e. the MSRV setting accurately reflects what we can compile with.