chore(largesmt): Fix the rocksdb build
The prior build configuration did not specify either of the bindgen
configurations (`bindgen-runtime` or `bindgen-static`) for binding to
`libclang`. This meant that the build could be inconsistent on some
systems (e.g. macOS developer machines).

We now specify `bindgen-runtime` which ensures that bindgen will
dynamically link to `libclang`. This has no performance impact as
`libclang` is only used in the `build.rs` build scripts and not in the
actual `miden-crypto` library.