Implement the history overlay mechanism (#679)
This mechanism is intended to be used for the forest implementation with
pluggable storage backends. This forest will eventually replace the
existing in-memory forest implementation, but for now is being developed
in a separate portion of the tree to avoid breakage. We are aiming to
merge portions of it incrementally to avoid mammoth PRs and increase
visibility, and this mechanism being relatively self-contained makes it
a good candidate for this.

This history mechanism is based on the following simplifying
assumptions:

- A forest is generally used to store multiple independent trees and
  versions of each tree, but does not need to share between those
  independent trees.
- Historical versions are not important to persist, and can hence be
  handled entirely in memory.

The commit contains the basic functionality of this history storage and
overlay mechanism, and tests that it functions correctly. It is almost
entirely internal code, with little change to the public interface, and
as a result temporarily uses `#[allow(dead_code)]` to keep the build
warnings clean. These will be removed in future commits.

The interface for the `History` in this commit is a best effort at the
design, and will likely need to undergo some minor changes to support
concurrency as the shape of said concurrency in the forest itself is
established.